refactor : [결제내역] 결제내역 목록과 상세 패키지 분리
