+++ presentation/src/main/java/com/nexters/boolti/presentation/reservationdetail/ResendGiftButton.kt
@@ -1,4 +1,4 @@
-package com.nexters.boolti.presentation.screen.reservations
+package com.nexters.boolti.presentation.reservationdetail
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
