+++ presentation/src/main/java/com/nexters/boolti/presentation/reservationdetail/ReservationDetailNavigation.kt
@@ -1,9 +1,8 @@
-package com.nexters.boolti.presentation.screen.reservations
+package com.nexters.boolti.presentation.reservationdetail
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.nexters.boolti.presentation.screen.MainDestination
-import com.nexters.boolti.presentation.screen.reservationId
 
 fun NavGraphBuilder.ReservationDetailScreen(
     navigateTo: (String) -> Unit,
